+++ public/vocables/modules/shared/ontologyModels.js
@@ -361,9 +361,9 @@ var OntologyModels = (function () {
                             /*if (!Config.sources[source] || !Config.topLevelOntologies[source]) {
                                 return callbackSeries();
                             }*/
-                                if (!Config.sources[source] ) {
-                                    return callbackSeries();
-                                }
+                            if (!Config.sources[source]) {
+                                return callbackSeries();
+                            }
                             var constraints = Config.ontologiesVocabularyModels[source].constraints;
                             for (var prop in Config.ontologiesVocabularyModels[source].properties) {
                                 var prop = Config.ontologiesVocabularyModels[source].properties[prop];