+++ public/vocables/modules/tools/KGquery/KGquery.js
@@ -327,9 +327,8 @@ var KGquery = (function () {
                 }
             }
 
-            if ( result.results) {
-                if (result.results.bindings.length == 0)
-                    return alert("no result");
+            if (result.results) {
+                if (result.results.bindings.length == 0) return alert("no result");
                 self.message("found items :" + result.results.bindings.length);
             }
 
@@ -341,10 +340,8 @@ var KGquery = (function () {
                 self.queryResultToTable(result);
             } else if (output == "Graph") {
                 self.queryResultToVisjsGraph(result);
-            }
-
-            else if (output == "shacl") {
-                KGconstraints_validator.process(result)
+            } else if (output == "shacl") {
+                KGconstraints_validator.process(result);
             } else {
                 Config.userTools.KGquery.toTools[output](result);
             }
@@ -508,20 +505,17 @@ var KGquery = (function () {
                         "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>" +
                         "PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>";
 
-
-                    var queryType="SELECT"
-                    if( options.output=="shacl"){
-                        queryType="CONSTRUCT"
-                        selectStr=""
+                    var queryType = "SELECT";
+                    if (options.output == "shacl") {
+                        queryType = "CONSTRUCT";
+                        selectStr = "";
                     }
-                    query += queryType+" " + selectStr + "  " + fromStr + " where {" + whereStr + "}";
+                    query += queryType + " " + selectStr + "  " + fromStr + " where {" + whereStr + "}";
 
                     query += " " + groupByStr + " limit 10000";
 
                     callbackSeries();
                 },
-
-
 
                 //execute query
                 function (callbackSeries) {
@@ -537,7 +531,7 @@ var KGquery = (function () {
                         query: query,
                         source: self.currentSource,
                     };
-                    query=Sparql_common. setPrefixesInSelectQuery(query)
+                    query = Sparql_common.setPrefixesInSelectQuery(query);
 
                     Sparql_proxy.querySPARQL_GET_proxy(url, query, "", { source: self.currentSource, caller: "getObjectRestrictions" }, function (err, result) {
                         if (err) {
@@ -822,19 +816,9 @@ var KGquery = (function () {
         }
     };
 
-    self.checkRequirements=function(){
-
-
-        KGquery.queryKG('shacl')
-
-
-
-
-    }
-
-
-
-
+    self.checkRequirements = function () {
+        KGquery.queryKG("shacl");
+    };
 
     return self;
 })();